requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
openai==1.59.6
pydantic==2.10.4
pypdf==5.1.0
python-docx==1.1.2
reportlab==4.2.5
python-dotenv==1.0.1
